spring-hw-10/src/main/java/ru/otus/spring/homework/oke/controller/AuthorController.java
package ru.otus.spring.homework.oke.controller;

import lombok.RequiredArgsConstructor;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;
import ru.otus.spring.homework.oke.dto.AuthorFullNameDto;
import ru.otus.spring.homework.oke.service.AuthorService;

import java.util.List;

@RestController
@RequiredArgsConstructor
public class AuthorController {
    private final AuthorService authorService;

    @GetMapping(value = "/api/author", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<AuthorFullNameDto> findAll() {
        return this.authorService.findAll();
    }
}
